fix: ChallengeListView->ParticipatingListView navigation link 오류 해결
